extract decorator kea testcase
